fix(topic): fetch topic detail page
